+++ clang/docs/ReleaseNotes.rst
@@ -901,6 +901,7 @@ OpenMP Support
 - Added support 'no_openmp_constructs' assumption clause.
 - Added support for 'self_maps' in map and requirement clause.
 - Added support for 'omp stripe' directive.
+- Added support for threadset clause in task and taskloop directives.
 
 Improvements
 ^^^^^^^^^^^^